// frontend/src/shared/api/__generated__/models/campaign.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keitaro Admin API
 * OpenAPI spec version: v1
 */
import type { CampaignType } from './campaignType';
import type { CampaignState } from './campaignState';
import type { CampaignCostType } from './campaignCostType';
import type { CampaignBindVisitors } from './campaignBindVisitors';
import type { SourceParameters } from './sourceParameters';
import type { S2SPostback } from './s2SPostback';

export interface Campaign {
  id?: number;
  /** Leave empty to generate automatically. */
  alias?: string;
  /** Set 'weight' to enable split-testing the flows. */
  type?: CampaignType;
  name?: string;
  cookies_ttl?: number;
  position?: number;
  state?: CampaignState;
  cost_type?: CampaignCostType;
  cost_value?: number;
  /** Leave empty to use tracker currency */
  cost_currency?: string;
  group_id?: number;
  /** 's' for flows. 'sl' for flows and landing pages. 'slo' for flows, landing pages and offers. */
  bind_visitors?: CampaignBindVisitors;
  traffic_source_id?: number;
  token?: string;
  cost_auto?: boolean;
  parameters?: SourceParameters;
  postbacks?: S2SPostback[];
  notes?: string;
  created_at?: string;
  updated_at?: string;
}